requirements.txt
asgiref==3.3.1
autopep8==1.5.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
et-xmlfile==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pkg-resources==0.0.0
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
